move tests around, move invalid cases to another file
