+++ src/aks-preview/azext_aks_preview/tests/latest/test_aks_commands.py
@@ -1,896 +1,3 @@
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_update_with_localdns_required_mode_one_valid_override_each(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        valid_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_with_valid_dns_overrides.json")
-        one_valid_override_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_one_valid_override_each.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "valid_config": valid_config_path,
-            "one_valid_override_config": one_valid_override_config_path
-        })
-
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={valid_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(add_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        update_cmd = (
-            "aks nodepool update --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --localdns-config={one_valid_override_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-        )
-        self.cmd(update_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        show_cmd = (
-            "aks nodepool show --resource-group={resource_group} --cluster-name={name} --name={nodepool_name}"
-        )
-        result = self.cmd(show_cmd).get_output_in_json()
-        assert result["localDnsProfile"]["mode"] == "Required"
-        # Validate that only one valid override each is present
-        assert len(result["localDnsProfile"].get("kubeDnsOverrides", {})) == 1
-        assert len(result["localDnsProfile"].get("vnetDnsOverrides", {})) == 1
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_update_with_localdns_invalid_mode(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        valid_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_with_valid_dns_overrides.json")
-        invalid_mode_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "invalid_mode.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "valid_config": valid_config_path,
-            "invalid_mode_config": invalid_mode_config_path
-        })
-
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={valid_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(add_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        update_cmd = (
-            "aks nodepool update --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --localdns-config={invalid_mode_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-        )
-        try:
-            self.cmd(update_cmd)
-            assert False, "Expected failure for invalid mode, but command succeeded."
-        except Exception as ex:
-            assert "invalid" in str(ex).lower() or "error" in str(ex).lower() or "mode" in str(ex).lower(), f"Unexpected error: {ex}"
-
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_update_with_localdns_null_config(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        valid_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_with_valid_dns_overrides.json")
-        null_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "null.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "valid_config": valid_config_path,
-            "null_config": null_config_path
-        })
-
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={valid_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(add_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        update_cmd = (
-            "aks nodepool update --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --localdns-config={null_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-        )
-        try:
-            self.cmd(update_cmd)
-            assert False, "Expected failure or default behavior for null config, but command succeeded."
-        except Exception as ex:
-            assert "invalid" in str(ex).lower() or "error" in str(ex).lower() or "null" in str(ex).lower(), f"Unexpected error: {ex}"
-
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_update_with_localdns_empty_config(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        valid_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_with_valid_dns_overrides.json")
-        empty_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "empty.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "valid_config": valid_config_path,
-            "empty_config": empty_config_path
-        })
-
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={valid_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(add_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        update_cmd = (
-            "aks nodepool update --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --localdns-config={empty_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-        )
-        try:
-            self.cmd(update_cmd)
-            assert False, "Expected failure or default behavior for empty config, but command succeeded."
-        except Exception as ex:
-            assert "invalid" in str(ex).lower() or "error" in str(ex).lower() or "empty" in str(ex).lower(), f"Unexpected error: {ex}"
-
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_update_with_localdns_required_mode_invalid_vnetdns(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        valid_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_with_valid_dns_overrides.json")
-        invalid_vnetdns_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_invalid_vnetdns.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "valid_config": valid_config_path,
-            "invalid_vnetdns_config": invalid_vnetdns_config_path
-        })
-
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={valid_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(add_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        update_cmd = (
-            "aks nodepool update --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --localdns-config={invalid_vnetdns_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-        )
-        try:
-            self.cmd(update_cmd)
-            assert False, "Expected failure due to invalid vnetDnsOverrides, but command succeeded."
-        except Exception as ex:
-            assert "invalid" in str(ex).lower() or "error" in str(ex).lower(), f"Unexpected error: {ex}"
-
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_update_with_localdns_required_mode_invalid_kubedns(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        valid_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_with_valid_dns_overrides.json")
-        invalid_kubedns_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_invalid_kubedns.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "valid_config": valid_config_path,
-            "invalid_kubedns_config": invalid_kubedns_config_path
-        })
-
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={valid_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(add_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        update_cmd = (
-            "aks nodepool update --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --localdns-config={invalid_kubedns_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-        )
-        try:
-            self.cmd(update_cmd)
-            assert False, "Expected failure due to invalid kubeDnsOverrides, but command succeeded."
-        except Exception as ex:
-            assert "invalid" in str(ex).lower() or "error" in str(ex).lower(), f"Unexpected error: {ex}"
-
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_add_with_localdns_missing_mode(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        missing_mode_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "missing_mode.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "missing_mode_config": missing_mode_config_path
-        })
-
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={missing_mode_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        try:
-            self.cmd(add_cmd)
-            assert False, "Expected failure or default behavior for missing mode, but command succeeded."
-        except Exception as ex:
-            assert "invalid" in str(ex).lower() or "error" in str(ex).lower() or "mode" in str(ex).lower(), f"Unexpected error: {ex}"
-
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_add_with_localdns_required_mode_empty_overrides(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        empty_overrides_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_empty_overrides.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "empty_overrides_config": empty_overrides_config_path
-        })
-
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={empty_overrides_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(add_cmd, checks=[self.check("provisioningState", "Succeeded")])
-        show_cmd = (
-            "aks nodepool show --resource-group={resource_group} --cluster-name={name} --name={nodepool_name}"
-        )
-        result = self.cmd(show_cmd).get_output_in_json()
-        assert result["localDnsProfile"]["mode"] == "Required"
-        assert result["localDnsProfile"].get("kubeDnsOverrides", {}) == {}
-        assert result["localDnsProfile"].get("vnetDnsOverrides", {}) == {}
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_add_with_localdns_required_mode_partial_invalid(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        partial_invalid_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_partial_invalid.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "partial_invalid_config": partial_invalid_config_path
-        })
-
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={partial_invalid_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        try:
-            self.cmd(add_cmd)
-            assert False, "Expected failure due to partial invalid override, but command succeeded."
-        except Exception as ex:
-            assert "invalid" in str(ex).lower() or "error" in str(ex).lower(), f"Unexpected error: {ex}"
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_add_with_localdns_required_mode_extra_property(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        extra_property_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_extra_property.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "extra_property_config": extra_property_config_path
-        })
-
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={extra_property_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(add_cmd, checks=[self.check("provisioningState", "Succeeded")])
-        show_cmd = (
-            "aks nodepool show --resource-group={resource_group} --cluster-name={name} --name={nodepool_name}"
-        )
-        result = self.cmd(show_cmd).get_output_in_json()
-        assert result["localDnsProfile"]["mode"] == "Required"
-        assert "extraProperty" not in result["localDnsProfile"]
-        assert_dns_overrides_equal(result["localDnsProfile"]["kubeDnsOverrides"], kubeDnsOverridesExpected)
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_add_with_localdns_required_mode_valid_vnetdns(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        vnetdns_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_with_valid_vnetdns.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "vnetdns_config": vnetdns_config_path
-        })
-
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={vnetdns_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(add_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        show_cmd = (
-            "aks nodepool show --resource-group={resource_group} --cluster-name={name} --name={nodepool_name}"
-        )
-        result = self.cmd(show_cmd).get_output_in_json()
-        assert result["localDnsProfile"]["mode"] == "Required"
-        assert "kubeDnsOverrides" not in result["localDnsProfile"] or result["localDnsProfile"]["kubeDnsOverrides"] is None
-        assert_dns_overrides_equal(result["localDnsProfile"]["vnetDnsOverrides"], vnetDnsOverridesExpected)
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_add_with_localdns_required_mode_valid_kubedns(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        kubedns_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_with_valid_kubedns.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "kubedns_config": kubedns_config_path
-        })
-
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={kubedns_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(add_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        show_cmd = (
-            "aks nodepool show --resource-group={resource_group} --cluster-name={name} --name={nodepool_name}"
-        )
-        result = self.cmd(show_cmd).get_output_in_json()
-        assert result["localDnsProfile"]["mode"] == "Required"
-        assert "vnetDnsOverrides" not in result["localDnsProfile"] or result["localDnsProfile"]["vnetDnsOverrides"] is None
-        assert_dns_overrides_equal(result["localDnsProfile"]["kubeDnsOverrides"], kubeDnsOverridesExpected)
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_add_with_localdns_empty_mode(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        empty_mode_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "empty_mode.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "empty_mode_config": empty_mode_config_path
-        })
-
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={empty_mode_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        try:
-            self.cmd(add_cmd)
-            assert False, "Expected failure or default behavior for empty mode, but command succeeded."
-        except Exception as ex:
-            assert "invalid" in str(ex).lower() or "error" in str(ex).lower() or "mode" in str(ex).lower() or "empty" in str(ex).lower(), f"Unexpected error: {ex}"
-
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_add_with_localdns_null_mode(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        null_mode_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "null_mode.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "null_mode_config": null_mode_config_path
-        })
-
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={null_mode_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        try:
-            self.cmd(add_cmd)
-            assert False, "Expected failure or default behavior for null mode, but command succeeded."
-        except Exception as ex:
-            assert "invalid" in str(ex).lower() or "error" in str(ex).lower() or "mode" in str(ex).lower() or "null" in str(ex).lower(), f"Unexpected error: {ex}"
-
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_add_with_localdns_empty_config(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        empty_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "empty.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "empty_config": empty_config_path
-        })
-
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={empty_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        try:
-            self.cmd(add_cmd)
-            assert False, "Expected failure or default behavior for empty config, but command succeeded."
-        except Exception as ex:
-            assert "invalid" in str(ex).lower() or "error" in str(ex).lower() or "empty" in str(ex).lower(), f"Unexpected error: {ex}"
-
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_add_with_localdns_null_config(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        null_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "null.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "null_config": null_config_path
-        })
-
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={null_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        try:
-            self.cmd(add_cmd)
-            assert False, "Expected failure or default behavior for null config, but command succeeded."
-        except Exception as ex:
-            assert "invalid" in str(ex).lower() or "error" in str(ex).lower() or "null" in str(ex).lower(), f"Unexpected error: {ex}"
-
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_add_with_localdns_invalid_mode(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        invalid_mode_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "invalid_mode.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "invalid_mode_config": invalid_mode_config_path
-        })
-
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={invalid_mode_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        try:
-            self.cmd(add_cmd)
-            assert False, "Expected failure for invalid mode, but command succeeded."
-        except Exception as ex:
-            assert "invalid" in str(ex).lower() or "error" in str(ex).lower() or "mode" in str(ex).lower(), f"Unexpected error: {ex}"
-
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_add_with_localdns_required_mode_invalid_vnetdns(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        invalid_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_invalid_vnetdns.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "invalid_config": invalid_config_path
-        })
-
-        # Create AKS cluster
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        # Add nodepool with required mode localdns config with invalid vnetDnsOverrides
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={invalid_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        try:
-            self.cmd(add_cmd)
-            assert False, "Expected failure due to invalid vnetDnsOverrides, but command succeeded."
-        except Exception as ex:
-            # Expecting an error due to invalid override
-            assert "invalid" in str(ex).lower() or "error" in str(ex).lower(), f"Unexpected error: {ex}"
-
-        # Clean up
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_add_with_localdns_required_mode_invalid_kubedns(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        invalid_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_invalid_kubedns.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "invalid_config": invalid_config_path
-        })
-
-        # Create AKS cluster
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        # Add nodepool with required mode localdns config with invalid kubeDnsOverrides
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={invalid_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        try:
-            self.cmd(add_cmd)
-            assert False, "Expected failure due to invalid kubeDnsOverrides, but command succeeded."
-        except Exception as ex:
-            # Expecting an error due to invalid override
-            assert "invalid" in str(ex).lower() or "error" in str(ex).lower(), f"Unexpected error: {ex}"
-
-        # Clean up
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_add_with_localdns_preferred_mode(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        preferred_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "preferred_mode_only.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "preferred_config": preferred_config_path
-        })
-
-        # Create AKS cluster
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        # Add nodepool with preferred mode localdns config
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={preferred_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(add_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        # Show nodepool and check localDNSProfile
-        show_cmd = (
-            "aks nodepool show --resource-group={resource_group} --cluster-name={name} --name={nodepool_name}"
-        )
-        result = self.cmd(show_cmd).get_output_in_json()
-        assert result["localDnsProfile"]["mode"] == "Preferred"
-        assert_dns_overrides_equal(result["localDnsProfile"]["kubeDnsOverrides"], kubeDnsOverridesExpectedDefault)
-        assert_dns_overrides_equal(result["localDnsProfile"]["vnetDnsOverrides"], vnetDnsOverridesExpectedDefault)
-        # Clean up
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
-
-    @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
-    def test_aks_nodepool_add_with_localdns_required_mode(self, resource_group, resource_group_location):
-        aks_name = self.create_random_name("cliakstest", 16)
-        nodepool_name = self.create_random_name("np", 6)
-        required_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_only.json")
-        self.kwargs.update({
-            "resource_group": resource_group,
-            "name": aks_name,
-            "nodepool_name": nodepool_name,
-            "ssh_key_value": self.generate_ssh_keys(),
-            "required_config": required_config_path
-        })
-
-        # Create AKS cluster
-        create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} "
-            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        # Add nodepool with required mode localdns config
-        add_cmd = (
-            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
-            "--name={nodepool_name} --node-count 1 --localdns-config={required_config} "
-            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
-            "--kubernetes-version 1.33.0"
-        )
-        self.cmd(add_cmd, checks=[self.check("provisioningState", "Succeeded")])
-
-        # Show nodepool and check localDNSProfile
-        show_cmd = (
-            "aks nodepool show --resource-group={resource_group} --cluster-name={name} --name={nodepool_name}"
-        )
-        result = self.cmd(show_cmd).get_output_in_json()
-    assert result["localDnsProfile"]["mode"] == "Required"
-    # Check for default DNS configuration (should not have kubeDnsOverrides or vnetDnsOverrides)
-    assert "kubeDnsOverrides" not in result["localDnsProfile"] or result["localDnsProfile"]["kubeDnsOverrides"] is None
-    assert "vnetDnsOverrides" not in result["localDnsProfile"] or result["localDnsProfile"]["vnetDnsOverrides"] is None
-        # Clean up
-        self.cmd(
-            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
-            checks=[self.is_empty()],
-        )
-
 # --------------------------------------------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
@@ -4327,6 +3434,126 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         assert_dns_overrides_equal(result["localDnsProfile"]["kubeDnsOverrides"], kubeDnsOverridesExpected)
         assert_dns_overrides_equal(result["localDnsProfile"]["vnetDnsOverrides"], vnetDnsOverridesExpected)
 
+        # Clean up
+        self.cmd(
+            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
+            checks=[self.is_empty()],
+        )
+
+    @AllowLargeResponse()
+    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
+    def test_aks_nodepool_update_with_localdns_required_mode_one_valid_override_each(self, resource_group, resource_group_location):
+        aks_name = self.create_random_name("cliakstest", 16)
+        nodepool_name = self.create_random_name("np", 6)
+        valid_dns_overrides_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_with_valid_dns_overrides.json")
+        kubedns_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_with_valid_kubedns.json")
+        vnetdns_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_with_valid_vnetdns.json")
+        self.kwargs.update({
+            "resource_group": resource_group,
+            "name": aks_name,
+            "nodepool_name": nodepool_name,
+            "ssh_key_value": self.generate_ssh_keys(),
+            "valid_dns_overrides": valid_dns_overrides_path,
+            "kubedns_config": kubedns_config_path,
+            "vnetdns_config": vnetdns_config_path
+        })
+
+        create_cmd = (
+            "aks create --resource-group={resource_group} --name={name} "
+            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
+            "--kubernetes-version 1.33.0"
+        )
+        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
+
+        # Add nodepool with valid DNS overrides
+        add_cmd = (
+            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
+            "--name={nodepool_name} --node-count 1 --localdns-config={valid_dns_overrides} "
+            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
+            "--kubernetes-version 1.33.0"
+        )
+        self.cmd(add_cmd, checks=[self.check("provisioningState", "Succeeded")])
+
+        show_cmd = (
+            "aks nodepool show --resource-group={resource_group} --cluster-name={name} --name={nodepool_name}"
+        )
+        result = self.cmd(show_cmd).get_output_in_json()
+        assert result["localDnsProfile"]["mode"] == "Required"
+        # Should have both kubeDnsOverrides and vnetDnsOverrides
+        assert_dns_overrides_equal(result["localDnsProfile"].get("kubeDnsOverrides", {}), kubeDnsOverridesExpected)
+        assert_dns_overrides_equal(result["localDnsProfile"].get("vnetDnsOverrides", {}), vnetDnsOverridesExpected)
+
+        # Update with only kubeDnsOverrides
+        update_kubedns_cmd = (
+            "aks nodepool update --resource-group={resource_group} --cluster-name={name} "
+            "--name={nodepool_name} --localdns-config={kubedns_config} "
+            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
+        )
+        self.cmd(update_kubedns_cmd, checks=[self.check("provisioningState", "Succeeded")])
+
+        result = self.cmd(show_cmd).get_output_in_json()
+        assert result["localDnsProfile"]["mode"] == "Required"
+        assert "vnetDnsOverrides" not in result["localDnsProfile"] or result["localDnsProfile"]["vnetDnsOverrides"] is None
+        assert_dns_overrides_equal(result["localDnsProfile"]["kubeDnsOverrides"], kubeDnsOverridesExpected)
+
+        # Update with only vnetDnsOverrides
+        update_vnetdns_cmd = (
+            "aks nodepool update --resource-group={resource_group} --cluster-name={name} "
+            "--name={nodepool_name} --localdns-config={vnetdns_config} "
+            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
+        )
+        self.cmd(update_vnetdns_cmd, checks=[self.check("provisioningState", "Succeeded")])
+
+        result = self.cmd(show_cmd).get_output_in_json()
+        assert result["localDnsProfile"]["mode"] == "Required"
+        assert "kubeDnsOverrides" not in result["localDnsProfile"] or result["localDnsProfile"]["kubeDnsOverrides"] is None
+        assert_dns_overrides_equal(result["localDnsProfile"]["vnetDnsOverrides"], vnetDnsOverridesExpected)
+
+        self.cmd(
+            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
+            checks=[self.is_empty()],
+        )
+
+    @AllowLargeResponse()
+    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
+    def test_aks_nodepool_add_with_localdns_required_mode(self, resource_group, resource_group_location):
+        aks_name = self.create_random_name("cliakstest", 16)
+        nodepool_name = self.create_random_name("np", 6)
+        required_config_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "localdnsconfig", "required_mode_only.json")
+        self.kwargs.update({
+            "resource_group": resource_group,
+            "name": aks_name,
+            "nodepool_name": nodepool_name,
+            "ssh_key_value": self.generate_ssh_keys(),
+            "required_config": required_config_path
+        })
+
+        # Create AKS cluster
+        create_cmd = (
+            "aks create --resource-group={resource_group} --name={name} "
+            "--node-count 1 --ssh-key-value={ssh_key_value} --generate-ssh-keys "
+            "--kubernetes-version 1.33.0"
+        )
+        self.cmd(create_cmd, checks=[self.check("provisioningState", "Succeeded")])
+
+        # Add nodepool with required mode localdns config
+        add_cmd = (
+            "aks nodepool add --resource-group={resource_group} --cluster-name={name} "
+            "--name={nodepool_name} --node-count 1 --localdns-config={required_config} "
+            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/LocalDNSPreview "
+            "--kubernetes-version 1.33.0"
+        )
+        self.cmd(add_cmd, checks=[self.check("provisioningState", "Succeeded")])
+
+        # Show nodepool and check localDNSProfile
+        show_cmd = (
+            "aks nodepool show --resource-group={resource_group} --cluster-name={name} --name={nodepool_name}"
+        )
+        result = self.cmd(show_cmd).get_output_in_json()
+        assert result["localDnsProfile"]["mode"] == "Required"
+        # Check for default DNS configuration (should not have kubeDnsOverrides or vnetDnsOverrides)
+        assert "kubeDnsOverrides" not in result["localDnsProfile"] or result["localDnsProfile"]["kubeDnsOverrides"] is None
+        assert "vnetDnsOverrides" not in result["localDnsProfile"] or result["localDnsProfile"]["vnetDnsOverrides"] is None
         # Clean up
         self.cmd(
             "aks delete --resource-group={resource_group} --name={name} --yes --no-wait",
